refacto: mapbox drone direction
